+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./home/schedule";
+export * from "./home/ranking";
+export * from "./home/revision";
+export * from "./location/location";
