+++ tsconfig.app.json
@@ -28,8 +28,11 @@
       "@routes/*": ["src/routes/*"],
       "@styles/*": ["src/styles/*"],
       "@utils/*": ["src/utils/*"],
+      "@assets/*": ["src/assets/*"],
       "@components/*": ["src/components/*"],
-      "@stores/*": ["src/stores/*"]
+      "@stores/*": ["src/stores/*"],
+      "@types/*": ["src/types/*"],
+      "@types": ["src/types"]
     }
   },
   "include": ["src"]
